Split up gui tests
Because one test is running es-mda and the other
an ensemble experiment with failing realizations
the results could be hard to track. Split it into
two files.
