+++ tests/unit_tests/gui/conftest.py
@@ -1,10 +1,20 @@
 import copy
+import fileinput
+import os.path
+import shutil
+import stat
 import time
 from datetime import datetime as dt
+from textwrap import dedent
+from typing import Tuple
 from unittest.mock import MagicMock, Mock
 
 import pytest
+from pytestqt.qtbot import QtBot
+from qtpy.QtCore import Qt, QTimer
+from qtpy.QtWidgets import QComboBox, QMessageBox, QWidget
 
+from ert._c_wrappers.enkf import EnKFMain, ErtConfig
 from ert.ensemble_evaluator.identifiers import CURRENT_MEMORY_USAGE, MAX_MEMORY_USAGE
 from ert.ensemble_evaluator.snapshot import (
     Job,
@@ -20,6 +30,183 @@ from ert.ensemble_evaluator.state import (
     REALIZATION_STATE_UNKNOWN,
     STEP_STATE_UNKNOWN,
 )
+from ert.gui.ertwidgets.caselist import AddRemoveWidget, CaseList
+from ert.gui.ertwidgets.closabledialog import ClosableDialog
+from ert.gui.ertwidgets.validateddialog import ValidatedDialog
+from ert.gui.main import GUILogHandler, _setup_main_window
+from ert.gui.simulation.run_dialog import RunDialog
+from ert.gui.simulation.simulation_panel import SimulationPanel
+from ert.gui.simulation.view import RealizationWidget
+from ert.gui.tools.manage_cases.case_init_configuration import (
+    CaseInitializationConfigurationPanel,
+)
+from ert.services import Storage
+from ert.shared.models import EnsembleExperiment, MultipleDataAssimilation
+
+
+def find_cases_dialog_and_panel(
+    gui, qtbot: QtBot
+) -> Tuple[ClosableDialog, CaseInitializationConfigurationPanel]:
+    qtbot.waitUntil(lambda: gui.findChild(ClosableDialog, "manage-cases") is not None)
+    dialog = gui.findChild(ClosableDialog, "manage-cases")
+    cases_panel = dialog.findChild(CaseInitializationConfigurationPanel)
+    assert isinstance(cases_panel, CaseInitializationConfigurationPanel)
+    return (dialog, cases_panel)
+
+
+@pytest.mark.usefixtures("use_tmpdir")
+@pytest.fixture(scope="module")
+def opened_main_window(source_root, tmpdir_factory):
+    with pytest.MonkeyPatch.context() as mp:
+        tmp_path = tmpdir_factory.mktemp("test-data")
+        shutil.copytree(
+            os.path.join(source_root, "test-data", "poly_example"),
+            tmp_path / "test_data",
+        )
+        mp.chdir(tmp_path / "test_data")
+        with fileinput.input("poly.ert", inplace=True) as fin:
+            for line in fin:
+                if "NUM_REALIZATIONS" in line:
+                    # Decrease the number of realizations to speed up the test,
+                    # if there is flakyness, this can be increased.
+                    print("NUM_REALIZATIONS 20", end="\n")
+                else:
+                    print(line, end="")
+            poly_case = EnKFMain(ErtConfig.from_file("poly.ert"))
+        args_mock = Mock()
+        args_mock.config = "poly.ert"
+
+        with Storage.init_service(
+            ert_config=args_mock.config,
+            project=os.path.abspath(poly_case.ert_config.ens_path),
+        ):
+            gui = _setup_main_window(poly_case, args_mock, GUILogHandler())
+            yield gui
+            gui.close()
+
+
+@pytest.mark.usefixtures("use_tmpdir, opened_main_window")
+@pytest.fixture(scope="module")
+def esmda_has_run(run_experiment):
+    # Runs a default ES-MDA run
+    run_experiment(MultipleDataAssimilation)
+
+
+@pytest.mark.usefixtures("use_tmpdir")
+@pytest.fixture(scope="module")
+def run_experiment(request, opened_main_window):
+    def func(experiment_mode):
+        qtbot = QtBot(request)
+        gui = opened_main_window
+        try:
+            shutil.rmtree("poly_out")
+        except FileNotFoundError:
+            pass
+        # Select correct experiment in the simulation panel
+        simulation_panel = gui.findChild(SimulationPanel)
+        assert isinstance(simulation_panel, SimulationPanel)
+        simulation_mode_combo = simulation_panel.findChild(QComboBox)
+        assert isinstance(simulation_mode_combo, QComboBox)
+        simulation_mode_combo.setCurrentText(experiment_mode.name())
+
+        # Click start simulation and agree to the message
+        start_simulation = simulation_panel.findChild(QWidget, name="start_simulation")
+
+        def handle_dialog():
+            message_box = gui.findChild(QMessageBox)
+            qtbot.mouseClick(message_box.buttons()[0], Qt.LeftButton)
+
+        QTimer.singleShot(500, handle_dialog)
+
+        # The Run dialog opens, click show details and wait until done appears
+        # then click it
+        def use_rundialog():
+            qtbot.waitUntil(lambda: gui.findChild(RunDialog) is not None)
+            run_dialog = gui.findChild(RunDialog)
+
+            qtbot.mouseClick(run_dialog.show_details_button, Qt.LeftButton)
+
+            qtbot.waitUntil(run_dialog.done_button.isVisible, timeout=100000)
+            qtbot.waitUntil(lambda: run_dialog._tab_widget.currentWidget() is not None)
+
+            # Assert that the number of boxes in the detailed view is
+            # equal to the number of realizations
+            realization_widget = run_dialog._tab_widget.currentWidget()
+            assert isinstance(realization_widget, RealizationWidget)
+            list_model = realization_widget._real_view.model()
+            assert list_model.rowCount() == simulation_panel.ert.getEnsembleSize()
+
+            qtbot.mouseClick(run_dialog.done_button, Qt.LeftButton)
+
+        QTimer.singleShot(1000, use_rundialog)
+        qtbot.mouseClick(start_simulation, Qt.LeftButton)
+
+    return func
+
+
+@pytest.mark.usefixtures("use_tmpdir")
+@pytest.fixture(scope="module")
+def ensemble_experiment_has_run(opened_main_window, run_experiment, request):
+    gui = opened_main_window
+    qtbot = QtBot(request)
+
+    def handle_dialog():
+        dialog, cases_panel = find_cases_dialog_and_panel(gui, qtbot)
+
+        # Open the create new cases tab
+        cases_panel.setCurrentIndex(0)
+        current_tab = cases_panel.currentWidget()
+        assert current_tab.objectName() == "create_new_case_tab"
+        create_widget = current_tab.findChild(AddRemoveWidget)
+        case_list = current_tab.findChild(CaseList)
+        assert isinstance(case_list, CaseList)
+
+        # Click add case and name it "iter-0"
+        def handle_add_dialog():
+            qtbot.waitUntil(lambda: current_tab.findChild(ValidatedDialog) is not None)
+            dialog = gui.findChild(ValidatedDialog)
+            dialog.param_name.setText("iter-0")
+            qtbot.mouseClick(dialog.ok_button, Qt.LeftButton)
+
+        QTimer.singleShot(1000, handle_add_dialog)
+        qtbot.mouseClick(create_widget.addButton, Qt.LeftButton)
+
+        dialog.close()
+
+    QTimer.singleShot(1000, handle_dialog)
+    manage_tool = gui.tools["Manage cases"]
+    manage_tool.trigger()
+
+    with open("poly_eval.py", "w", encoding="utf-8") as f:
+        f.write(
+            dedent(
+                """#!/usr/bin/env python
+import numpy as np
+import sys
+import json
+
+def _load_coeffs(filename):
+    with open(filename, encoding="utf-8") as f:
+        return json.load(f)
+
+def _evaluate(coeffs, x):
+    return coeffs["a"] * x**2 + coeffs["b"] * x + coeffs["c"]
+
+if __name__ == "__main__":
+    if np.random.random(1) > 0.5:
+        sys.exit(1)
+    coeffs = _load_coeffs("coeffs.json")
+    output = [_evaluate(coeffs, x) for x in range(10)]
+    with open("poly_0.out", "w", encoding="utf-8") as f:
+        f.write("\\n".join(map(str, output)))
+        """
+            )
+        )
+    os.chmod(
+        "poly_eval.py",
+        os.stat("poly_eval.py").st_mode | stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH,
+    )
+    run_experiment(EnsembleExperiment)
 
 
 @pytest.fixture()
